chip: npcx: add some ugly guards for zephyr compilation
coreboot-sdk takes less nicely to some of the chip/ files being used
than zephyr-sdk did:

  https://paste.googleplex.com/5737573268848640

Generally speaking, we should never allow compiling code in chip/ like
this in the future to prevent things like this.  But given the damage
was already done, this is the least-invasive way to move past it.

BUG=b:175248959
BRANCH=none
TEST=compile both with zephyr-sdk and coreboot-sdk for volteer

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I3a25c12e890d89f5a4ba67d3ed5eb2ec3280b720