zephyr: Update the I2C table
Something has happened since this code was written and now it doesn't
work. The current code checks the status of each of the named-i2c-ports
nodes, which always seems to return 0, then ends up searching for a
device called "".

Obviously this code did work and should work, but I'm uploading this
just because it makes the battery command work for me:

On the other hand, I don't think we need to check the status. After
all, we are search for a device by its node label and if the node is
disabled it will not have a device anyway.

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I709421be86027e0e6166a7658fe59ba68783cbe7