ide-support: Add vscode default settings.json and gitignores
This adds a .default settings.json file for vscode, so that you
do not need to copy/paste form the ide-support.md doc. This also
makes it easier for people that don't known that the ide-support.md
doc exists.

This also adds the .vscode directory to the gitignores, so that local
customization/settings are not always in the git status.

BRANCH=none
BUG=none
TEST=Checked in Gitiles
TEST=Check that my settings.json is not in "git status".
TEST=Check that ec.code-workspace is not in "git status".

Change-Id: I4dd8cfcd79c291a6f4fb0d573df131f0cb0d9844
Signed-off-by: Craig Hesling <hesling@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2586041
Commit-Queue: Jett Rink <jettrink@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>