quiche: Adding tcpmv2.0 support and SRC mode working
This CL enables usb-pd support for quiche and includes support to
establish a PD contract. It does not include any alt-dp mode support
although those config options are added in order to make the compile
work.

This CL also replaces the virtual mux with the PS8822 usb_mux driver.

BUG=b:167601672
BRANCH=None
TEST=Verify that PD contract is established on quiche

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Idd0185698ba641e7dfaba028ebb33e9f1054eb1f

quiche: Add support for ps8822 demux

This CL replaces the virtual mux with the PS8822 usb_mux driver. The
hpd_update routine that's included is not implemented at this point
in the stack.

BUG=b:175660576
BRANCH=None
TEST=Connect on host port and then read back mode register to confirm
that it's been set as expected.

Change-Id: Id2ca25b84e937fa4f66c0b61ccb9b327110d704f