lazor: Add motion sensors ICM40608 and KX022
KX022 is for lid accel. ICM40608 is for base accel and gyro.
For the i2c addresses of KX022 and BMA255 are different, do
a single read to check which chip is on the DUT.
For the i2c addresses of ICM40608 and BMI160 are the same,
read reg ICM426XX_REG_WHO_AM_I to check if the chip is
ICM426XX_CHIP_ICM40608.

BUG=b:175663607
BRANCH=none
TEST=Motion sensors are detected correctly and the motion
     function works.

Signed-off-by: Sue Chen <sue.chen@quanta.corp-partner.google.com>
Change-Id: I2788bdc1e5bad1b1212ffaa7670969318b77221f