bb_retimer: Update retimer init sequence
Retimer init sequence should follow below:
1. Deassert RT_RST, sleep for 1ms
2. Send I2C message to initialize the sequence
3. If NAK from above, sleep for 40ms/20ms
   and resend same I2C
4. If ACK from step 2, skip step 3.

The nvm exclusive access was to wait for RESET_N to deassert, but since
the RESET_N lines are deasserted at same time this is handled in SPI
flash HW logic we no longer require it. Also the mutex was introduced to
reduce the delay from 40ms to 20ms, but since the delay is after the
RESET_N lines, it is redundant as well.

BUG=b:174240235
BRANCH=None
TEST=Tested with TBT dock, enumeration on
 both ports are fine and swapping between ports also
 works fine.

Change-Id: I6cf2a49af6c1ab463821448df92d9f9a8ca5041e
Signed-off-by: Divya Sasidharan <divya.s.sasidharan@intel.com>
Signed-off-by: Ayushee Shah <ayushee.shah@intel.com>