pompom: Move Refresh key to row 3
Starting from pompom rev2, the refresh key has been moved to F2, which
maps to row 3 (KSI3) in the keyboard matrix.

This CL intends to work together with a hardware change - routing KSI3
to H1, which will be implemented in the next pompom hardware spin.

In pompom rev2 boards, since KSI2 is still routed to H1, this CL
won't magically make the three finger gesture `Esc+Refresh+Pwr` work.
We have to live with `Esc+F3+Pwr` temporarily in rev2 boards.

Note that this CL will break the non-H1-triggered recovery with rev2.
Since this is not what a normal user will do, I think it is OK to land
the CL now.

BUG=b:174290322
BRANCH=trogdor
TEST=Entering recovery mode via `Esc+F3+Pwr` still works in pompom rev2

Signed-off-by: Philip Chen <philipchen@google.com>
Change-Id: I9a3376e18a800d1023994d264fbed0eb0a30fdb9