TCPMv2: Update source-out configs
Now that the DPM will be handling source-out decisions for TCPMv2,
remove references to its old configuration options from TCPMv2 boards in
order to avoid any confusion as to what code is running now.  Also
remove the charge manager notifications of sink attach/detach since the
policy is being centralized into the DPM.

Note that the previous configuration options only ever allocated one 3.0
A port, and so the default number of 3.0 A ports has been set to 1.

BRANCH=None
BUG=b:168862110,b:141690755
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ie452e3da32b04226503539daa67b6b9f4a58aa58