npcx9: remove unnecessary default CONFIGs for internal flash
In npcx9, the internal flash is accessed via FIU/UMA (implementing in
chip/npcx/flash.c). It won’t use the common/spi_flash.c to access the
internal flash. This CL remove the two default configs related to the
SPI flash in the npcx9 chip config and fix the npcx9_evb setting for
the SPI flash.

BRANCH=none
BUG=b:165777478
TEST=buildall pass

Signed-off-by: Wealian Liao <whliao@nuvoton.corp-partner.google.com>
Signed-off-by: CHLin <CHLin56@nuvoton.com>
Change-Id: I7c525791d001fa2833b7d895f277768948783938