TCPMv2: Update current limits in appropriate task
Since setting the current limit on a port often involves communicating
with a port's PPC or charger, run this from the appropriate port task
rather than allowing another port to run it.  This interface can also be
used later to increase or decrease the source Rp for non-PD partners.

BRANCH=None
BUG=b:168862110,b:141690755
TEST=on drawcia, plug in two hubs requiring 3.0 A and confirm that the
first receives 3.0 A, and on its unplug the second is offered 3.0 A

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ie8423f612d199f4389955e00a82e9ca588f2d836