tcpm: Move tcpm.h into an include directory
This header file is used from quite a few files, relying on the EC
build system to find includes in the driver/tcpm directory. For Zephyr
we don't want to add that as an include.

It makes more sense for header files to be in an include directory, so
move it and fix up the users.

BUG=b:175434113
BRANCH=none
TEST=build Zephyr and ECOS on volteer

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I5851914b1a7d3fdc1ba911c0fbe9046afbaf6f5d