TCPMv2: Add common hpd->dp_atten converter
This CL adds a new file for UFP_D ports. The inclusion of this file is
controlled via a new config option CONFIG_USB_PD_ALT_MODE_UFP. Note
that this is not intended to be used with TCPMv1 UFP_D legacy devices.

The intitial version of this file includes the hpd to usb pd
DP_ATTENTION message converter. The implementaiton follows the spec in
terms of both states and hpd event queueing rules.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I5c7f4a3d226eb7b33553c90f04c69c3d80bd8672