USB: Remove DTS identification from USB PD polarity
We've added differentiation for DTS polarity and normal
device polarity. These values were being passed to the
pd[port].polarity fields. This creates problems for the
usb_pd_config.h files which operate on the assumption
that polarity is a boolean value of 0 or 1.

In the case of ternary blocks that follow the format of:
'polarity ? A | B', if polarity is 'POLARITY_CC1_DTS = 2'
we are incorrectly identifying it as flipped.

Other usages like Servo_v4 which have lookup arrays like:
tx_gpio[port][polarity]. In these cases, the DTS polarities
point to garbage memory and generate hard-faults.

BRANCH=none
BUG=b:176154334
TEST=Connect to servo console and run 'cc snkdts'
    Validate that the servo did not hard-fault, instead
    loads a valid configuration.

Signed-off-by: Brian Nemec <bnemec@chromium.org>
Change-Id: If5ed079fee9e91dbf291e1f210762c298948ad1e