sasuke: Rerun USB-C interrupts if still asserted
This commit revises the USB-C interrupt handlers.  Sometimes the
interrupt line can remain asserted once we've finished handling the
interrupt.  This commit leverages the work that was done for the other
dedede boards to check if the interrupt is still asserted and rerun
the interrupt handlers if needed.

BUG=b:143166332
BRANCH=dedede
TEST=`make -j BOARD=sasuke`

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: Ia3fcd1f38e6de10259276d171128fb5dfa967f47