usb_mux: Optimize reading mux state on unconnected port
If Type-C port has nothing connected then the query
of mux state should not power on the mux for reading
the state.
This helps reduce unnecessary power on/off retimer
or mux for ports with no device attached to it.

BUG=b:174240235
BRANCH=None
TEST=Test enumeration of type-c devices is not affected.
Also check on voxel if retimer init is not called periodically
when device is not connected.

Change-Id: I2f1c4093889164c0d47c3ac33e7a46ab524bcce0
Signed-off-by: Divya Sasidharan <divya.s.sasidharan@intel.com>