Docs: Update TCPMv2 source policy to include FRS
FRS ports which require 3A should also be considered in the source-out
policy.  This updates the diagram to include FRS and includes FRS in the
example flow.  Additionally, it changes the policy for allocation from
"oldest port" to "lowest number port" since that policy requires less
flash space.

BRANCH=None
BUG=b:141690755
TEST=verify markdown renders in Chrome with Markdown Preview Plus

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I14e359696fa7dc15ae349cade6fc723e7998d941