zephyr: Rename the top-level Kconfig
Update PLATFORM_EC to refer to the shim since that is really what it is
brining in. Once we actually have code on the Zephyr side then perhaps
we can introduce a distinction here, but that is not the current goal.

Also use Chromium OS EC consistently.

BUG=b:176449230
BRANCH=none
TEST=ninja -C /tmp/z/vol/build-ro menuconfig
See that the naming appears

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I11aefa5ba224b22a5bb91c4a4ed146ad5b0c500c