TCPMv2: Decrease Low Power Mode debounce delay to 35 ms
This patch decreases PD_LPM_DEBOUNCE to 35 ms. With current
100 ms delay ANX7447 can't go to Low Power Mode due to silicon
bug (b:77544959). When DRP mode is enabled ANX7447 triggers alert
every 39.7 ms. Checking what caused an alert invloves I2C
transaction which changes timeout after which TCPC goes to Low
Power Mode to get_time().val + PD_LPM_DEBOUNCE_US. As a result it
is not possible to put TCPC to LPM which is necessary to work
properly (b:149761477)

Original value for PD_LPM_DEBOUNCE_US was 10 ms, introduced
in CL:1119255. This value is still used in TCPMv1.
For TCPMv2, PD_LPM_DEBOUNCE_US was changed to 100 ms in CL:2126078.
It is unclear why this value was changed. Review discussion and bug
b:149772936 don't mention why we needed to change that.

Issue was found when running EC ToT with TCPMv2 enabled on bobba board.

BUG=b:161775827, b:77544959, b:149761477
BRANCH=none
TEST=Flash EC ToT on octopus board with ANX7447 (eg. bobba).
     Make sure that DRP mode is enabled. Connect charger to port
     with ANX7447 chip. Make sure that EC can connect with charger.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ia08818378a1fca424d392b4699f9c4adc460bfb5