zephyr: Connect HAS_TASK_CHARGER and PLATFORM_EC_BATTERY
These options are linked in that we cannot have one without
the other.

Select the charger task automatically if PLATFORM_EC_BATTERY is
enabled, since that feature cannot work without the task. If
PLATFORM_EC_BATTERY is not enabled there is no need for the task.

BUG=b:176449230
BRANCH=none
TEST=ninja -C /tmp/z/vol/build-ro menuconfig
Check the operation of HAS_TASK_CHARGER and PLATFORM_EC_BATTERY

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I14ada8716053649ac13ea12a7b06c2ae1702df42