zephyr: Update naming and help for tasks
The name is currently too verbose and the help is too brief. Update to
make this easier to read and more useful.

Drop the task name from 'Stack size' since it is implied and just makes
things harder to follow.

Also put the tasks into alphabetical order.

BUG=b:176449230
BRANCH=none
TEST=ninja -C /tmp/z/vol/build-ro menuconfig
See that the new task options appear with help

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I7514cc730e1a33c60388eb4a0f0ce2c800056499