make/util: Move utils make var init to board/build.mk
We have tons of boards in EC now. We should be a bit more selective
and less hackey when selecting which utils/artifacts are needed for
a given board.

One such odd example is that stm32mon is compiled for all boards.
Another example is that servo_ectool is only build for nocturne_fp,
but the logic to control that was hard coded into util/build.mk.

This shift the initialization/control to the board and baseboard's
build.mk and allows the util/build.mk to add items to these variables.

Ideally, we would use multiple variables to aggregate the utils selection
from the three sources (baseboard/build.mk, board/build.mk, and
util/build.mk), but I don't think the addition complexity is currently
warranted.

BRANCH=none
BUG=none
TEST=some

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: I23153850eda10fc1c88d386b9f3cd8296df9c6f1