Put charger headers into their own directory
These header file is used from quite a few files, relying on the EC
build system to find includes in the driver/charger directory. For
Zephyr we don't want to add that as an include.

It makes more sense for header files to be in an include directory, so
move it and fix up the users.

BUG=b:175434113
BRANCH=none
TEST=make BOARD=volteer -j30
build volteer on zephyr

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I5b810f53cdf545a885f3977849f9f2ca1d04d60a