Split out some USB-C config into new files
The Zephyr build needs this information but cannot use the whole
board.c file. It is pretty big anyway, so let's move some of the USB-C
config info into a separate file.

Add a header file for the part specific to this config.

We can progressively move more as Zephyr is able to build it. This
requires the addition of many Kconfig options, so will be done in
stages.

Sadly other volteer boards need to follow the same approach, since the
baseboard code needs to be able to include a header file to get the
definition for 'enum usbc_port'. They are all the same apart from
halvor. So this means a new header file for each board using the
volteer baseboard. Perhaps there is a better way?

BUG=b:175434113
BRANCH=none
TEST=make -j30 buildall
With a zephyr-chrome CL, build volteer on zephyr
Check for other boards:
for b in $(grep volteer board/*/build.mk |awk -F/ '{print $2}'); do \
    echo $b; grep -A5 "enum usbc_port" board/$b/board.h; done

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I2a4bb0af2b8571ce992cdbbc8362dec25176872e