Split out some USB-C config into new files
The Zephyr build needs this information but cannot use the whole
board.c file. It is pretty big anyway, so let's move some of the USB-C
config info into a separate file.

Add a header file for the part specific to this config.

We can progressively move more as Zephyr is able to build it. This
requires the addition of many Kconfig options, so will be done in
stages.

BUG=b:175434113
BRANCH=none
TEST=make BOARD=volteer -j30  (why doesn't -s work?)
With a zephyr-chrome CL, build volteer on zephyr

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I2a4bb0af2b8571ce992cdbbc8362dec25176872e