Create public headers for a few PPC/TCPM drivers
At present boards includes the private header of some of the drivers.
This is not ideal but it works. For Zephyr we don't really want to
access headers in private driver directories.

Instead, create public headers for the five drivers needed by the
Zephyr volteer build. For now, include the public header in the
private header (the one included by the EC code), so that fewer code
changes are required.

BUG=b:175434113
BRANCH=none
TEST=make buildall -j30   (way too verbose to see what is happening)
build volteer on zephyr

Change-Id: I5b810f53cdf545a885f3977849f9f2ca1d04d60a
Signed-off-by: Simon Glass <sjg@chromium.org>