Move tcpic.h header into include/driver
This header cannot currently be accessed by Zephyr since it is in a
driver directory, not an include directory. This header has quite a
bit of public stuff in it, so it seems reasonable to consider
everything public.

Move the header file and update all users.

BUG=b:175434113
BRANCH=none
TEST=make buildall -j30
build volteer on zephyr

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ibba37f47a06783fafb5095f853f2a68d92b6df87