u2f: Add authorization_secret_hash as a field in versioned kh
The purpose of the authorization secret hash is to enforce the
correct authorization secret in the case where presence is waived.
The authorization_secret_hash is one input material to the
authorization_hmac.

Previous implementation doesn't include the authorization secret
hash in the u2f_versioned_key_handle struct, and that's misleading,
so we add authorization_secret_hash field. A unittest is added for
this change.

A u2fd change must be made to ignore the changes in u2f.h, otherwise
generating versioned key handle will be broken until the next cr50
firmware uprev that includes this change.

BUG=b:172971998
TEST=make run-u2f
TEST=make CR50_DEV=1 BOARD=cr50

Change-Id: I06c25f0831aeeedf7b5a5ac39011c1769fa8a38b
Signed-off-by: Yicheng Li <yichengli@chromium.org>