zephyr: add a fake shared memory implementation for posix-ec
CL:2575207 was not tested against posix-ec and caused build issues.

posix-ec does not have the _image_ram_end symbol available.  This adds
a new configuration for a fake 1MB shared memory option, and enables
it for posix architecture by default.

Note that if this CL (or an alternative fix) cannot be landed in a
timely fashion, CL:2575207 and any CLs which depend on it should be
reverted so we can enable the CQ again and stop stepping on our own
toes.

BUG=b:176828988
BRANCH=none
TEST=run zephyr-chrome firmware_builder.py

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I1d828f480c6a5da0b162a124c8a5a62ae8afb444