usb_charger: Support CONFIG to make VBUS supplier as default
In the usb_charger, the VBUS supplier (5V 0.5A) is used as default
before the Type-C detection or the BC 1.2 detection. However, it
causes an issue on some ARM boards and it needs to be disabled.

So add the new CONFIG_VBUS_SUPPLIER_AS_DEFAULT, which is defined
by default.

The issue happens on the first boot from battery cutoff, the EC
RO detects 5V 3A and boots AP. When sysjump to RW, this VBUS
supplier (5V 0.5A) is used first and limits the charger input
current to some lower value. It makes AP out of power.

If the power adapter just supports 5V 0.5A, the BC 1.2 will detect
it a SDP and add the SDP supplier. The VBUS supplier is actually
not needed.

BRANCH=Trogdor
BUG=b:174105232, b:175663604
TEST=Undefined the CONFIG, verified the first boot from cutoff,
and checked the Type-C detection and the BC 1.2 detection.

Change-Id: Ief75c5ac6e1ea8421137f5a4c2a0aea4981bbb30
Signed-off-by: Wai-Hong Tam <waihong@google.com>