Trogdor: Do not use VBUS supplier as default
Undefine the CONFIG_VBUS_SUPPLIER_AS_DEFAULT.

It fixes the issue which happens on the first boot from battery
cutoff, the EC RO detects 5V 3A and boots AP. When sysjump to RW,
the VBUS supplier (5V 0.5A) is used first and limits the charger
input current to some lower value. It makes AP out of power.

If the power adapter just supports 5V 0.5A, the BC 1.2 will detect
it a SDP and add the SDP supplier. The VBUS supplier is actually
not needed.

BRANCH=Trogdor
BUG=b:174105232, b:175663604
TEST=Undefined the CONFIG, verified the first boot from cutoff,
and checked the Type-C detection and the BC 1.2 detection.

Change-Id: Ibad10ee20887bf73e9f182f09608a7c6a23874f7
Signed-off-by: Wai-Hong Tam <waihong@google.com>