SYV682x: Improve Sink OCP Handling
Previously the sink OCP was calling the source OCP handler, which is
incorrect. Remove that, and replace it with equivalent messages and
retries in the driver. since TSD was previously treated the same as OCP,
split that into its own message.

There is no PD significance to a sink OCP, so just re-enable the sink
path 3 times. The 4rd time disables the sink path and requires a device
reconnect. This will make us more robust to non-compliant sources.

BUG=b:175444003
TEST=Short PPVAR_VBUS_IN to GND with a 2 ohm resistor on Voxel while
charging at 15V. See that OCP is reported 4 times, then disabled.
TEST=make buildall
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: Id5eb04819f1e9e2001e3e68283da09a812faeff3